Extract guided terms help modal
